pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisprly"
version = "0.1.0"
description = "A voice-controlled automation tool"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "plfavreau", email = "pierrelouis.favreau@gmail.com"},
]
maintainers = [
    {name = "plfavreau", email = "pierrelouis.favreau@gmail.com"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "groq==0.29.0",
    "keyboard==0.13.5",
    "numpy==2.3.1",
    "PyQt6==6.9.1",
    "pyautogui==0.9.54",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "scipy==1.16.0",
    "sounddevice==0.5.2",
    "soundfile==0.13.1",
    "pyinstaller",
    "pillow",
    "psutil>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/plfavreau/whisprly"
Repository = "https://github.com/plfavreau/whisprly"
Issues = "https://github.com/plfavreau/whisprly/issues"

[project.scripts]
whisprly = "whisprly.main:main"

[tool.setuptools]
packages = ["whisprly"]

[tool.setuptools.package-data]
whisprly = ["*.txt", "*.md"]
